requirements.txt
# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
einops>=0.7.0

# Computer Vision
opencv-python>=4.8.0
albumentations>=1.3.0
facenet-pytorch>=2.5.0

# Data & Visualization
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Experiment Tracking
wandb>=0.15.0
tensorboard>=2.14.0

# Utilities
scikit-learn>=1.3.0
tqdm>=4.66.0
hydra-core>=1.3.0
pytest>=7.4.0
pillow>=10.0.0